Rework default values for Country and Language
With this change, only actually existant translations are shown in
the selection menu for the languages. In order to provide a default
value if the user's phone is set to some Locale for which no actual
translation exists, provide a sensible fallback mechanism.

For example, if the phone is set to de_AT, this looks for an exact
match for 'de_AT' first. This is not successful, so it tries 'de'
next and matches Locale.GERMAN, which will be preselected.

If no match at all is found (for example, something like dk_DK),
we fall back to Locale.ENGLISH.

With those changes, ConfigurationUtils.getDefaultCountryCode()
could return the empty string (as the Locale in the Configuration
might not contain a country. In this case, use the phone's
setting as the default value.